[feat] Move Slider components' styled elements to separate file
